Handlers y lógica de delete card

+++ VipManagement.API/Controllers/CardController.cs
@@ -8,55 +8,51 @@ using MediatR;
 
 namespace VipManagement.API.Controllers
 {
-
     [ApiController]
     [Route("api/[controller]")]
     public class CardController : ControllerBase
     {
-        // Llegeix el dbcontext
         private readonly VipManagementDbContext _context;
+        private readonly IMediator _mediator;
 
-
-
-        public CardController(VipManagementDbContext context)
+        public CardController(VipManagementDbContext context, IMediator mediator)
         {
-            // es crea una copia del contexte actual
             _context = context;
+            _mediator = mediator;
         }
 
         [HttpGet]
-
         public async Task<IActionResult> Get()
         {
-            var Cards = await _context.Cards.ToListAsync();
-            return Ok(Cards);
+            var cards = await _context.Cards.ToListAsync();
+            return Ok(cards);
         }
 
-
-
         [HttpPost("createCard")]
-        public async Task<ActionResult> Create([FromBody]CreateCardInputDto dto)
+        public async Task<ActionResult> Create([FromBody] CreateCardInputDto dto)
         {
             var cmd = (CreateCardCommand)dto;
 
-            var result = await Mediator.Send(cmd);
+            var result = await _mediator.Send(cmd);
+
             return Ok(result);
         }
 
-        [HttpDelete]
-        public async Task<IActionResult> Delete (Card card)
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            var card = await _context.Cards
+                .FirstOrDefaultAsync(x => x.Id == new CardId(id));
+
+            if (card is null)
+            {
+                return NotFound();
+            }
+
             _context.Cards.Remove(card);
             await _context.SaveChangesAsync();
+
             return NoContent();
-            
-
-          
         }
-
-
-
-
-       
     }
 }
